adapt to python3

+++ dataset_loaders/composite.py
@@ -41,13 +41,13 @@ class MF(data.Dataset):
     self.no_duplicates = no_duplicates
 
     if dataset == '7Scenes':
-      from seven_scenes import SevenScenes
+      from .seven_scenes import SevenScenes
       self.dset = SevenScenes(*args, real=self.real, **kwargs)
       if self.include_vos and self.real:
         self.gt_dset = SevenScenes(*args, skip_images=True, real=False,
           **kwargs)
     elif dataset == 'RobotCar':
-      from robotcar import RobotCar
+      from .robotcar import RobotCar
       self.dset = RobotCar(*args, real=self.real, **kwargs)
       if self.include_vos and self.real:
         self.gt_dset = RobotCar(*args, skip_images=True, real=False,
@@ -63,7 +63,7 @@ class MF(data.Dataset):
     else:
       skips = self.skip * np.ones(self.steps-1)
     offsets = np.insert(skips, 0, 0).cumsum()
-    offsets -= offsets[len(offsets) / 2]
+    offsets -= offsets[int(len(offsets) / 2)]
     if self.no_duplicates:
       offsets += self.steps/2 * self.skip
     offsets = offsets.astype(np.int)
@@ -135,11 +135,11 @@ class OnlyPoses(data.Dataset):
   def __init__(self, dataset, *args, **kwargs):
     kwargs = dict(kwargs, skip_images=True)
     if dataset == '7Scenes':
-      from seven_scenes import SevenScenes
+      from .seven_scenes import SevenScenes
       self.real_dset = SevenScenes(*args, real=True, **kwargs)
       self.gt_dset   = SevenScenes(*args, real=False, **kwargs)
     elif dataset == 'RobotCar':
-      from robotcar import RobotCar
+      from .robotcar import RobotCar
       self.real_dset = RobotCar(*args, real=True, **kwargs)
       self.gt_dset   = RobotCar(*args, real=False, **kwargs)
     else:
